Made bg tile invisible
Made bg tile invisible, so it would not reduce performance (on fill-rate limited systems). Still keep it as an object to make single tile pickable.  V99 of ThreeJS should solve issue with horizon (buildings are white in BG), and restore correct color for tile floors.

Note: The cache is not really useful here. But I would like to merge this PR to solve the performance issue, and solve the cache issue later.