Remove faulty test-npm-packages.sh check
The script assumes that the repository versions and the versions in
npmjs.com match, which is not always true, for example after we bumped
the versions in preparation for publishing.

Since this is blocking publishing, we need to remove it until we find a
more correct solution to this problem.