HARP-6099: Subdivide tiles in the world space of their tiling scheme.
This change generalizes the subdividion modifier so it can be used
to subdivide geometries in their world space.

Subdividing in the world space of the data ensures that no geometry
crossing the tile boundaries is introduced while subdividing, also
it makes possible to generate texture coordinates for the new vertices.

Signed-off-by: Roberto Raggi <roberto.raggi@here.com>