HARP-5722: Fix TileTreeTraverse.subTiles() for half quadtree scheme.
SubTiles.length is always 1 << (2*level), independently from the mask passed
on the construction. TileTreeTraverse.subTiles() was using that length member
to loop through all child tiles. For half quadtree subdivision scheme however,
the root tile has 2 children, instead of 4, so 2 of the 4 children returned where
invalid.

To fix this, SubTiles is now an iterable and the child tiles can be obtained
by iterating with a for..of loop.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>