HARP-6782: Use the requested $layer to speed up StyleSetEvaluator.
Add a naive visitor that searches for usages of
`["==", ["get", "$layer"], "some layer name"]` inside ["all, e1, ...eN]
expressions. When such usages are found the requested layer name
is stored in the `_layer`  member of `StyleInternalParams` and it used
to filter the evaluation of the conditions that need to be checked
by `getMatchingTechniques`.

Also, compile the style conditions when the StyleSetEvaluator is
constructed instead of lazily compile the expressions before
evaluating them.

Signed-off-by: Roberto Raggi <roberto.raggi@here.com>