HARP-6270: Frustum evaluation takes highest buildings into account.
Because we do not yet store nor calculate the buildings elevations in
the VisibleTileSet, we need to account for tallest building height
(on earth).
This extends highest elevation read from tile set with the Dubai's tower
Burj Khalifa height, giving rough but always correct assumption where the
near plane needs to be.
Patch-set fixes the problem with tall buildings clipping when camera is
very close ground surface.

Signed-off-by: Krystian Kostecki <ext-krystian.kostecki@here.com>