HARP:6270: Calculate visible tiles set min/max elevation and update frustum.
This approach stores min/max elevation in TileKeyEntry and then in Tile
objects itself, this allows to store elevation in both cached and new tiles.
Later on only visible tiles are parsed to calculate min/max elevation
renderable and this information is used to update ClipPlanesEvaluator
and take terrain features into account when evaluating new frustum.

Signed-off-by: Krystian Kostecki <ext-krystian.kostecki@here.com>