HARP-7270: Remove maxVisibleDataSourceTiles property from MapViewOptions.
Since it is difficult to predict the number of tiles to be displayed cause
it depends on geo-projection used and camera pose, we decided to remove
this constraint (limit of tiles) entirely. The number of tiles should
be anyway limited by proper frustum culling and correct near/far frustum
planes calculation (see ClipPlanesEvaluator).

Signed-off-by: Krystian Kostecki <ext-krystian.kostecki@here.com>