HARP-7356: Fixed wrong handling of tiles that only contain text elements but no geometry
If a tile contains no geometry but just text the Tile.hasGeometry flag was set to false (b/c Tile.forceGeometry was not called). Since hasGeometry is used in the VisibleTileSet computation to deduce if a tile is loaded(what is also totally wrong but will be adressed in a separate PR) no tile was rendered.
With this change forceGeometry will be called if the tile has no geometry but text.
Additionally the original decodedTile will be kept and it will not be replaced with a dummy empty decodedTile (this was also requested by OLP)