HARP-7270: Change maxVisibleDataSourceTiles use from MapViewOptions.
Since it is difficult to predict the number of tiles to be displayed cause
it depends on geo-projection used and camera pose, we decided to set this
constraint (limit of tiles) to extreme value and not relate it to cache
size.
The number of tiles should be anyway limited by proper frustum culling and
correct near/far frustum planes calculation (see ClipPlanesEvaluator).

Signed-off-by: Krystian Kostecki <ext-krystian.kostecki@here.com>