HARP-7373: Temporary fix, recover overlay text from text renderer on reset.
When text element renderer is reset once the theme is loaded, the overlay
text added to it was being lost. Add temporary change that moves the overlay
text to the newly text renderer instance.
Incoming PR will change the way the text renderer is initialized so that
an instance is created once on MapView constructor and never deleted.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>