HARP-7808 Fallback tiles have lower render order.
This ensures that the parent tiles don't overlap other surrounding
child tiles, which is especially problematic for borders with
transparency.

Note, this introduces a regression, which will be fixed by introducing
clipping planes.

Also, the BackgroundDataSource needs to have a lower renderOrder than the
groundPlane of the tiles, otherwise the fallback tiles would be covered
by the BackgroundDataSource, effectively disabling fallback.

Change-Id: Ia42d81637d3c220861be7cc039329bb8bf2b2c32
Signed-off-by: Jonathan Stichbury <2533428+nzjony@users.noreply.github.com>