HARP-7808 Adding clip planes for the Tiles.
- Clipping is done in the pixel shader, this introduces 4 dot products
hence this should be tested for peformance.

- They aren't configurable in the style because this would require
runtime recompilation of the shaders, big performance impact.

- If there are users who complain about the performance of this, we could
use the stencil buffer, however this works only for planar projections.

- We don't just enable clipping on the neighbours, because this is a non
trivial algorithm. It could be added in future (we do this on the native
side to avoid skirts for the terrain tiles where the LOD doesn't differ)

Change-Id: I9c41faaa27ff915771f9f76b014e9f277ff69ed2
Signed-off-by: Jonathan Stichbury <2533428+nzjony@users.noreply.github.com>