HARP-8066 Fixed bug with checkedTiles
checkedTiles was always set when a tile was looked at, unfortunately this
was also used to early exit, to prevent the recursion for sibling tiles
unfortunately, this early exit was incorrect, because we didn't know if
a parent tile was actually found, just that it was checked. Now we store
if the parentTile was found, which means we can then search for children
after checked the parent was checked.

Signed-off-by: Jonathan Stichbury <2533428+nzjony@users.noreply.github.com>