HARP-7373: Remove unnecessary render batch check in PoiRenderer.computeIconScreenBox()
A precondition to call computeIconScreenBox() is that prepareRender() was successfully
called first. If that's true, then poiInfo.poiRenderBatch should have the assigned a valid batch
index. That's checked within an assertion. There's no further need to double check that the batch
can be retrieved, and there's no need for the method to return a boolean.
The method is now static since it doesn't use any object member.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>