HARP-7373: Allocate screen space for label only if it's not rejected.
This is a minor improvement that avoids flickering in a particular scenario:

Imagine we are looking at "EUROPE" label, then we zoom in to a level where continent
labels are disabled and capitals become visible. "London" starts fading in to the
west side (it doesn't collide with "EUROPE") as "EUROPE" keeps fading out. As soon
as "EUROPE" becomes invisible, the label "DEUTSCHLAND" that was in collision both
with "EUROPE" and "London" starts fading in, since it has less priority than "EUROPE"
but more than "London", and causes the rejection of "London" that starts fading out.
If we let "DEUTSCHLAND" fade in as soon as "EUROPE" starts fading out, then "London"
will never fade in in the first place.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>