HARP-8361: Fix unnecessary materials & uniforms updates in MapMeshMaterials.
MapMeshMaterials defines mixins that are injected into standard THREE.js
materials to allow adding new features:
- fading,
- extrusion,
- displacement map.

The logic of mixin features was very inefficient and time consuming, such
as it forced many materials (and uniform) updates (re-compilation).
This patchset limits such WebGl shader re-compilation and uniforms
transfer to minimum. It also provides the chaining of onBeforeCompile
callback so all the mixins may be applied simultaneously. Previously there
was a bug that allowed only one feature to be applied to THREE.Material.
TODO:
- possibly implement generic mixins that could be applied also to
EdgeMaterial and SolidLineMaterial.

Signed-off-by: Krystian Kostecki <ext-krystian.kostecki@here.com>