UrlUtils: URI reference resolution cross-platform
In order to determine if a path of a referenced resource in a scheme (or child
URI) is absolute, the string is tested against a regular expression which only
covered non-Windows paths. This led to problems when actually working with
resource paths on Windows.

- Added unit test

This change allows to detect absolute paths in a Windows-format as well

Signed-off-by: ESTIB <jose.vega@here.com>