MINOR: Remove 3D objects from root more efficiently.
At the beginning of a frame, the previous frame objects are
removed from the scene calling MapView's mapTilesRoot.remove()
for each one of them. Each remove() call searches for the object
in an array and then splices it.
It's much simpler to just set the root's children array to length 0.

The average speedup is small, but it gets rid of some peaks where the
whole removal was taking ~1ms.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>