HARP-8971: Fix wrong computation of fullFrameTime (#1307)
Previously the fullFrameTime was computed by subtracting the last frame start time from the current frame start time.
This was done to make sure we also measure the time we spend outside the current animation frame (i.e. geometry creation time).
Unfortunately with this approach we also measure everything the application/test is doing.

With this change the fullFrameTime is explicitly computed by adding the geometry creation time to the current frame.