HARP-9344: WebTileDataSource - don't use const enums.
`const enum`s are not emitted to Javascript by default, so javascript
users would see `WebTileDataSource.ppiValue` as `undefined`.

Note `const enum`s are there only for performance reasons and those
contants are not intended to be used in performance sensitive code so
there is no reason to hack it around with `preserveConstEnums` compiler
flag.

Signed-off-by: Zbigniew Zagorski <ext-zbyszek.zagorski@here.com>