Add support for filtering features based on their state.
This change adds APIs to manipulate the state of a feature.
Note, that currently the feature state can only be used to
enable/disable features. That is, harp.gl themes can use
`["feature-state", property]` in expressions assigned to
the `enabled` attribute, for example:

    {
        when: ["==", ["geometry-type"], "LineString"],
        layer: "roads",
        technique: "solid-line",
        renderOrder: 10000,
        attr: {
            enabled: ["feature-state", "enabled"],
            lineWidth: "1px",
            color: "#c00"
        }
    }

Application code can programmatically update the state of a feature
using the DataSource API, for example:

    dataSource.setFeatureState(featureId, {
        enabled: true
    });
    mapView.update()

Signed-off-by: Roberto Raggi <roberto.raggi@here.com>