HARP-9275: Refactor handling of tile bounding boxes.
- DecodedTile may provide either:
a) A bounding box (e.g. for pure 3D geometry). The tile will use that bounding box,
no elevation will be applied to it, and it's geo box will be computed to match it.

b) The maximum Geometry height in meters. The tile will compute the bounding box from
its geo box. They resulting bbox may be elevated (if terrain data available) and will be
extended to include the maximum geometry height.

- minElevation and maxElevation properties are now private.

- maxGeometryHeight getter/setter removed.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>