HARP-9275: Test intersection with tiles on picking. (#1379)
* HARP-9275: Test intersection with tiles on picking.

Only objects in intersected tiles are tested for intersection.

Tile bounding boxes are updated to go from minElevation to
maxElevation + maxGeometryHeight.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>

* HARP-9275: Address review comments.

- Remove unnecessary MapView update in PickHandler.
- Add unit tests to Tile for BBox update.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>

* HARP-9275: Refactor handling of tile bounding boxes.

- DecodedTile may provide either:
a) A bounding box (e.g. for pure 3D geometry). The tile will use that bounding box,
no elevation will be applied to it, and it's geo box will be computed to match it.

b) The maximum Geometry height in meters. The tile will compute the bounding box from
its geo box. They resulting bbox may be elevated (if terrain data available) and will be
extended to include the maximum geometry height.

- minElevation and maxElevation properties are now private.

- maxGeometryHeight getter/setter removed.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>

* HARP-9275: Address review comments and fix TileGeometryCreatorTest.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>