HARP-9275: Implement VertexCache for mesh on the fly transformations.
It'll be used for mesh displacements on picking. Terrain overlay is done
by displacing meshes in the vertex shader. To pick those meshes, they
are displaced in CPU. To avoid allocating the whole position buffer again,
intersection test is done on the fly for every triangle after it's displaced.

To avoid recomputations of the same vertex, a small LRU cache is used.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>