HARP-9898: Change the way how text offset is applied to the Labels.
Changes the way the x/yOffset label's text offsets are applied, so text
displacement is absolute and always in the same screen space direction,
regardless of the text alignment / placement and its fits the unified
styling and native implementation.

Recent implementation of text offsets attributes defined in theme (xOffset,
yOffset) was to move the label's text by specified number of screen
pixels away from label origin or label's icon edge. The direction of this
movement (offset) was directly related with the text alignment. For
`hAlignment==Left` and `vAlignment==Above` the offset was simply added to
actual screen space position, but for `Right` and `Below` alignments
the corresponding offsets were subtracted. So the offsets worked like
a padding (or margin), always applied in direction of text placement.

Patchset also fixes the ScreenCollisions debug draw to be accurate and
display also candidates for screen allocations (not overlapping anymore).

Signed-off-by: Krystian Kostecki <ext-krystian.kostecki@here.com>