HARP-10478: Apply text POIs margin both to collision box and text offset.
Additional text margin (used as padding between icon and text) was not
applied to the text position, resulting in growing collision box, but
without effect on relative distance between label icon and text.
This path fixes it and adds support to bounding box measurement via
TextCanvas that are not 0, 0 anchored. This means that change will work
even without margin applied.

Signed-off-by: Krystian Kostecki <ext-krystian.kostecki@here.com>