HARP-10496: Add support for feature-space texture coords.
This change introduces basic support for computing texture
coordinates in feature space. When a technique requests
feature-space texture coordinates the decoder uses
the `bbox` property of the feature, defined as the tuple
`[west, south, east, north]` to calculate the uv coordinates.

Also, this change adds a simple example demonstrating how
to use the feature-space texture coordinates with a custom
shader technique.

Signed-off-by: Roberto Raggi <roberto.raggi@here.com>