HARP-5644: Support for Political Views via FeatureModifiers stack. (#1570)
* HARP-5644: Support for Political Views via FeatureModfier.

When different political view is requested, special implementation of
FeatureModifier is initialized that converts border lines "kind"
attributes according to different point of view, i.e. re-writing generic
"kind" with a contents of "kind:xx", where "xx" denotes country code.
This results in assigning new kinds'v values:
- unrecognized_disputed,
- unrecognized_country,
to generic "kind". Of course this kinds require separate/unique styling
policy, because current themes does not define any styling for it.

From now on features modifiers where are processed in queue, one
after another. It allows to add virtually unlimited number of data 
modifiers to the list while the result of first modifier is the input of
the next one.

Signed-off-by: Krystian Kostecki <ext-krystian.kostecki@here.com>