HARP-10446: Animate newly loaded tiles.
There's only one single animation ongoing at each moment in time. Any newly
loaded tiles while the animation is ongoing will start animating with the current
extrusion ratio. Once the animation is finished, if a new tile (that doesn't belong
to the same subtree of any of the finished tiles) is added, a new animation is started,
which will apply to any new tile, leaving the finished tiles unchanged.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>