HARP-10446: Simplify extruded building animation. (#1600)
* HARP-10446: Simplify extruded building animation.

Animation is done now using a unique extrusion ratio for all tiles. Animation
starts when the view's zoom level is greater or equal to the minimum zoom level
for extruded buildings.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>

* HARP-10446: Animate newly loaded tiles.

There's only one single animation ongoing at each moment in time. Any newly
loaded tiles while the animation is ongoing will start animating with the current
extrusion ratio. Once the animation is finished, if a new tile (that doesn't belong
to the same subtree of any of the finished tiles) is added, a new animation is started,
which will apply to any new tile, leaving the finished tiles unchanged.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>

* HARP-10446: Use AnimationExtrusionHandler.enabled if nothing is specified
on the style.

Also disable extrusion animation for geojson IBCT tests.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>

* HARP-10446: Change how new tiles are animated.

For a new tile, check if any ancestor or descendant was already animated,
skip animation in that case.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>