HARP-10765 Ensure 1x1 mipmap texture generated... (#1664)
for non square textures (i.e. when the padded width or height is different).

I.e. a texture 60x40 would be padded to 64x64, so there would be no issue, but a texture 60x32 would
be padded to 64x32 and we would miss generating the final 1x1 texture (the final texture would have
been 2x1), which causes an error in WebGL:

RENDER WARNING: texture bound to texture unit 0 is not renderable. It might be non-power-of-2 or
have incompatible texture filtering (maybe)?

Adding the final 1x1 mipmap removes this warning and the icon renders.

Signed-off-by: Jonathan Stichbury <2533428+nzjony@users.noreply.github.com>