Add environment variable HARP_NO_HARD_SOURCE_CACHE.
The `HARP_NO_HARD_SOURCE_CACHE` env variable can be used to
control the usage of `hard-source-webpack-plugin`. For example:

    HARP_NO_HARD_SOURCE_CACHE=true yarn start

Also, this change disables the usage of the hard-source cache
in the CI.

Note that the value assigned to `HARP_NO_HARD_SOURCE_CACHE` is
not important, that is, as long as the environment variable
`HARP_NO_HARD_SOURCE_CACHE` is defined the hard-source cache will
be disabled.

Signed-off-by: Roberto Raggi <roberto.raggi@here.com>