HARP-11212: Fix undefined behavior when computing the edge color.
The previous code was mixing vertex colors with the edge color
to compute the final color value. The mix was assuming that the
geometry was using vertex colors. This is not the case for the
majority of our themes.

This change introduces the parameter `verteColors` and
the property `objectColor` that can be used to correctly
compute the final edge color when vertex colors are not
available.

Signed-off-by: Roberto Raggi <roberto.raggi@here.com>