HARP-11152: Fix extruded polygon outlines when style has no minZoomLevel.
Extrusion animation is enabled by default, when a extruded polygon style
does not define a value "animateExtrusion" value. In this case, if no
minZoomLevel is specified, extrusion does not start until zoom level 16
by default, and therefore outlines are not rendered.

To fix it, extrusion animation is disabled if neither "animateExtrusion"
nor "minZoomLevel" are defined in the style.