HARP-11894 Fixing partially missing tiles (#2040)
The problem was that the perspective division looses information about
whether the tile was behind the camera or behind the far plane, which
means we can't sort the tiles correctly. Tiles behind the near plane
are therefore given a distance of -1 and those past the far plane
are given a distance of +1. Note, we know that the tiles are visible
because they intersect the frustum.

Signed-off-by: Jonathan Stichbury <2533428+nzjony@users.noreply.github.com>