HARP-13812: Tune the default value of extruded-polygon's constantHeight.
This change makes the default value of the `constantHeight`
property of the `extruded-polygon` technique depend on the tile's
storage level. That is, unless explictly set, the projection scaling
factor used to convert meters to world units is applied
only if the tile's level is greater than or equal to 12. This
change of behavior may results to a different height when rendering
extruded polygons coming from tiles at low storage levels but it
will avoid possible artifacts when rendering large polygons.

Signed-off-by: Roberto Raggi <roberto.raggi@here.com>