HARP-14470 Move from nyc to karma
nyc only supports code coverage in node, however we have a number of browser
based tests, and it is the primary use case, so we have decided to do the
code coverage in karma.

Note, the bundled package from webpack isn't enough because we need the
separate files to be shown in the HTML coverage report, hence we need to
configure the karma.conf.js which files to test / transpile and which
can just be served locally (i.e. resources).

Signed-off-by: Jonathan Stichbury <2533428+nzjony@users.noreply.github.com>