HARP-14678 Fix issue with morton codes being stored as 32 bit
The tile dependency example wasn't working with two separate data sources because the morton code
was stored in a Float32Array, which doesn't have enough precision and caused issues. This change
stores the row, column and level separately and then in the decoder, the mortonCode is computed and
stored as a `number`, which has more accuracy that a 32 bit float.

Signed-off-by: Jonathan Stichbury <2533428+nzjony@users.noreply.github.com>