HARP-16397 Don't run the test-npm-packages.sh script
when releasing, because the script gets confused, it installs dependencies from the current latest package but
then installs the new packages and the peer dependency mismatch causes issues.

I'm not sure if this is what caused the issue as mentioned in the ticket HARP-16397.

Anyway, we need to merge back the release branch to master, and at that point this script will run, so worst case, we can
fix it by doing a patch release.

Signed-off-by: Jonathan Stichbury <2533428+nzjony@users.noreply.github.com>