MAPSJS-2660: Disable depth test for IconMaterial.
Depth buffer is not cleared after rendering 3D scene. As a results
icons were not passing the depth test in some situations, like in
globe with top/down view.
Depth test is not needed, labels are always on top of other objects
and they don't write on the depth buffer.

Signed-off-by: Andres Mandado <andres.mandado-almajano@here.com>