MAPSJS-2983 Write to the console if we need too many attempts getting the image from the cache
Some failed attempts are ok, because there are cases where the DataSource is attached and ready
and the map has started to render, but the theme isn't fully set.

There are other ways to fix the problem, but they are generally more invasive and could introduce
regressions / performance issues, some options are:
- In the TileGeometryCreator::createAllGeometries method, await on the getTheme() Promise
- Don't allow new TextElements to be added while the theme is updating
- Investigate and remove all calls to update the mapview while the method `addDataSource` is called
  and see if this can be removed, because no update calls means we don't render and request new tiles
  which may contain references to POI's that don't yet exist on the cache.

For the sake of simplicity, this is the preferred method.

Change-Id: I441a0ff5f63fd3d2a403b43b422c3044c902c7a6
Signed-off-by: Jonathan Stichbury <2533428+nzjony@users.noreply.github.com>